post and fetch comments. reload page fixed.
right now if you want to delete comments, (on mongodb compass or web) you need to delete both in the post's comments array collection and in the comments collection.
for reloading page is done. by doing store the user data on the AuthContext.js 's frontend.
mistake note: the comments path's frontend will have to include /comments/ at the first as same as which app.use() in the app.js